requirements.txt
streamlit>=1.20.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=1.13.1
googletrans==4.0.0-rc1
gTTS>=2.3.0
